refactor(tests): split telegram renderer and dedupe FakeCommand
test_renderer.py was 178 LOC mixing text/render-many behaviour with
five media-type suites. Pull TestImage / TestVideo / TestAudio /
TestSticker / TestRaw into test_renderer_media.py (123 LOC); leave
TestText and TestRenderMany in test_renderer.py (44 LOC). Move the
renderer fixture and render() helper to conftest.py.

test_bot.py was 178 LOC because it carried its own 33 LOC FakeCommand
class. Move FakeCommand into telegram/conftest.py so any future
test_bot variant or sibling can reuse it; test_bot.py drops to 137 LOC.
